last commit

+++ src/RestTraining/Domain/Repositories/BoundedPeriodRepository.cs
@@ -1,10 +1,13 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
 using System.Linq;
 using System.Linq.Expressions;
+using RestTraining.Api.DTO;
 using RestTraining.Api.Domain.Entities;
 using RestTraining.Api.Domain.Services;
+using RestTraining.Api.Infrastructure;
 
 namespace RestTraining.Api.Domain.Repositories
 {
@@ -43,10 +46,17 @@ namespace RestTraining.Api.Domain.Repositories
 
         public void InsertOrUpdate(BoundedPeriod boundedPeriod)
         {
-            if (!_bookingDatesService.IsBoundedPeriodValid(_context, boundedPeriod))
+            if (boundedPeriod.Id == default(int))
             {
-                throw new ArgumentException("This period intersectы with another one already existed", "BeginDate, EndDate");
+                PreInsertCheck(boundedPeriod);
             }
+            else
+            {
+                PreUpdateCheck(boundedPeriod);
+            } 
+            
+            DatesValidCheck(boundedPeriod);
+            
             if (boundedPeriod.Id == default(int))
             {
                 _context.BoundedPeriods.Add(boundedPeriod);
@@ -56,16 +66,64 @@ namespace RestTraining.Api.Domain.Repositories
                 _context.Entry(boundedPeriod).State = EntityState.Modified;
             }
         }
-        
+
+        private void DatesValidCheck(BoundedPeriod boundedPeriod)
+        {
+            if (!_bookingDatesService.IsBoundedPeriodValid(_context, boundedPeriod))
+            {
+                throw new BoundedPeriodDatesException();
+            }
+        }
+
+        private void PreUpdateCheck(BoundedPeriod boundedPeriod)
+        {
+            PreInsertCheck(boundedPeriod);
+            if (_context.BoundedPeriods.Find(boundedPeriod.Id) == null)
+            {
+                throw new ParameterNotFoundException();
+            }
+        }
+
+        private void PreInsertCheck(BoundedPeriod boundedPeriod)
+        {
+            if (_context.BoundedReservationsHotels.Find(boundedPeriod.BoundedReservationsHotelId) == null)
+            {
+                throw new ParameterNotFoundException();
+            }
+        }
+
         public void Delete(int id)
         {
             var boundedPeriod = _context.BoundedPeriods.Find(id);
+            if (boundedPeriod == null)
+            {
+                throw new ParameterNotFoundException();
+            }
             _context.BoundedPeriods.Remove(boundedPeriod);
         }
 
         public void Save()
         {
             _context.SaveChanges();
+        }
+
+        public List<BoundedPeriod> GetAllForHotel(int hotelId)
+        {
+            CheckHotelIsValid(hotelId);
+            return All.Where(x => x.BoundedReservationsHotelId == hotelId).ToList();
+        }
+
+        public BoundedPeriod GetByHotelIdAndId(int hotelId, int id)
+        {
+            CheckHotelIsValid(hotelId);
+            return All.SingleOrDefault(x => x.BoundedReservationsHotelId == hotelId && x.Id == id);
+        }
+
+        private void CheckHotelIsValid(int hotelId)
+        {
+            var hotel = _context.BoundedReservationsHotels.Find(hotelId);
+            if (hotel == null)
+                throw new ParameterNotFoundException();
         }
 
         public void Dispose()
